refactor: move build qualified name helper to base class
